Rename Hlothhaere and Eadric to Hlothhaere-and-Eadric.md
Summary: The Kingdom of Kent, known as Cantaware Rīce in Old English and Regnum Cantuariorum in Latin, was an early medieval kingdom in what is now South East England. It was established in the fifth or sixth century CE and continued to exist until being fully absorbed into the Kingdom of England in the tenth century.